requirements.txt
langchain_community
python-dotenv
streamlit
langchain
groq
faiss-cpu
huggingface_hub
bs4
sentence-transformers
pypdf
transformers
fpdf
python-docx
pymupdf
pdfkit
requests
pdfminer.six
cryptography
fpdf